+++ frontend/src/backend/wails/trigger.ts
@@ -2,12 +2,14 @@ import { createEvent } from "effector";
 
 export const onExitDirty = createEvent<any>();
 export const onShowError = createEvent<string>();
+export const onUpgradeAvailable = createEvent<string>();
 
 export const spleFrontError = createEvent<void>();
 
 const enum FrontEventType {
   ExitDirty = "exitDirty",
   ShowError = "showError",
+  UpgradeAvailable = "upgradeAvailable",
 }
 interface FrontEvent<K extends FrontEventType, V = null> {
   type: K;
@@ -15,7 +17,8 @@ interface FrontEvent<K extends FrontEventType, V = null> {
 }
 type FrontEvents =
   | FrontEvent<FrontEventType.ExitDirty>
-  | FrontEvent<FrontEventType.ShowError, string>;
+  | FrontEvent<FrontEventType.ShowError, string>
+  | FrontEvent<FrontEventType.UpgradeAvailable, string>;
 
 export const spleFrontEvent = (e: FrontEvents) => {
   switch (e.type) {
@@ -25,6 +28,10 @@ export const spleFrontEvent = (e: FrontEvents) => {
 
     case FrontEventType.ShowError:
       onShowError(e.data);
+      return;
+
+    case FrontEventType.UpgradeAvailable:
+      onUpgradeAvailable(e.data);
       return;
   }
 };